[FAWRY-29] refactor: optimize outbox event processing with chunking and transaction management
- Replaced bulk fetch with chunked retrieval (findTop10ByProcessed) to avoid memory issues
- Introduced programmatic transaction management using TransactionTemplate
- Marked processed outbox events instead of deleting, for better traceability and resilience
- Ensured proper logging and structured while-loop to handle reprocessing of unprocessed events
- Improved separation of concerns by moving logic to a dedicated ScheduledOutboxEventProcessor class
